* remove some compiler warnings relate to domains * move gsExpressions, gsExprHelper * move expressions out of gsExpressions
