+++ src/gsDomain/gsDomainIterator.h
@@ -417,7 +417,10 @@ public:
     { GISMO_ERROR("Cannot proceed to next element. End iterator reached."); }
 
     virtual void next(index_t increment) override
-    { GISMO_ERROR("Cannot proceed to next element. End iterator reached."); }
+    {
+        GISMO_UNUSED(increment);
+        GISMO_ERROR("Cannot proceed to next element. End iterator reached.");
+    }
 
     virtual void prev() override
     {
@@ -426,6 +429,7 @@ public:
 
     virtual void prev(index_t decrement) override
     {
+        GISMO_UNUSED(decrement);
         GISMO_NO_IMPLEMENTATION
     }
