+++ src/gsDomain/gsHDomainIterator.h
@@ -61,7 +61,6 @@ public:
     {
         m_leaf = this->init(m_tree);
         updateLeaf();
-        updateElement();
     }
 
 
@@ -97,14 +96,8 @@ public:
     {
         bool isGood = nextLexicographic(m_curElement, m_meshStart, m_meshEnd);
 
-        if (isGood) // new element in m_leaf
-            updateElement();
-        else // went through all elements in m_leaf
-        {
+        if (!isGood) // new element in m_leaf
             isGood = nextLeaf();
-            if (isGood)
-                updateElement();
-        }
     }
 
     // ---> Documentation in gsDomainIterator.h
@@ -121,9 +114,6 @@ public:
             if (!isGood)
                 isGood = nextLeaf();
         }
-
-        if (isGood)
-            updateElement();
     }
 
     /// Resets the iterator so that it can be used for another
@@ -132,7 +122,6 @@ public:
     {
         m_leaf = m_tree.beginLeafIterator();
         updateLeaf();
-        updateElement();
     }
 
     gsVector<T> lowerCorner() const override
@@ -227,14 +216,6 @@ private:
             // for n breaks, we have n - 1 elements (spans)
             m_meshEnd(dim) =  m_breaks[dim].end() - 1;
         }
-    }
-
-    /// Computes lower, upper and center point of the current element, maps the reference
-    /// quadrature nodes and weights to the current element, and computes the
-    /// active functions.
-    GISMO_DEPRECATED
-    void updateElement()
-    {
     }
 
 // =============================================================================
