+++ src/gsExpressions/gsExpressions.h
@@ -14,9 +14,7 @@
 #pragma once
 
 #include <gsCore/gsFuncData.h>
-#include <gsAssembler/gsDirichletValues.h>
 #include <gsMSplines/gsMappedBasis.h>
-
 
 namespace gismo
 {
@@ -59,38 +57,7 @@ void secDerToHessian(const gsEigen::DenseBase<Derived> &  secDers,
     }
 }
 
-/// Struct containing information for matrix assembly
-template<class T> struct gsFeSpaceData
-{
-    gsFeSpaceData(const gsFunctionSet<T> & _fs, index_t _dim, index_t _id):
-    fs(&_fs), dim(give(_dim)), id(give(_id)) { }
 
-    const gsFunctionSet<T> * fs;
-    index_t dim, id;
-    gsDofMapper mapper;
-    gsMatrix<T> fixedDofs;
-    index_t cont; //int. coupling
-
-    bool valid() const
-    {
-        GISMO_ASSERT(nullptr!=fs, "Invalid pointer.");
-        return static_cast<size_t>(fs->size()*dim)==mapper.mapSize();
-    }
-
-    void init()
-    {
-        GISMO_ASSERT(nullptr!=fs, "Invalid pointer.");
-        if (const gsMultiBasis<T> * mb =
-            dynamic_cast<const gsMultiBasis<T>*>(fs) )
-            mapper = gsDofMapper(*mb, dim );
-        else if (const gsBasis<T> * b =
-                 dynamic_cast<const gsBasis<T>*>(fs) )
-            mapper = gsDofMapper(*b, dim );
-        mapper.finalize();
-        fixedDofs.clear();
-        cont = -1;
-    }
-};
 
 // Forward declaration in gismo namespace
 template<class T> class gsExprHelper;
@@ -110,8 +77,12 @@ template <> struct is_arithmetic<real_t>{enum{value=1};};
 template <typename E, bool = is_arithmetic<E>::value >
 class _expr {using E::GISMO_ERROR_expr;};
 
+template<class E> class symbol_expr;
+template<class T> class gsNullExpr;
 template<class T> class gsFeSpace;
+template<class T> class gsFeElement;
 template<class T> class gsFeVariable;
+template<class T> class gsComposition;
 template<class T> class gsFeSolution;
 template<class E> class symm_expr;
 template<class E> class symmetrize_expr;
@@ -385,192 +356,26 @@ std::ostream &operator<<(std::ostream &os, const _expr<E> & b)
 }
 }
 
-#include <gsAssembler/expr/precomputed_expr.h>
+#include <gsExpressions/precomputed_expr.h>
+
+#include <gsExpressions/gsFeSpace.h>
+#include <gsExpressions/gsFeVariable.h>
+#include <gsExpressions/gsFeSolution.h>
+#include <gsExpressions/gsFeElement.h>
+#include <gsExpressions/gsComposition.h>
+
+#include <gsExpressions/gsNullExpr.h>
+#include <gsExpressions/gsGeometryMap.h>
+
+#include <gsExpressions/symbol_expr.h>
+#include <gsExpressions/integral_expr.h>
+#include <gsExpressions/col_expr.h>
+#include <gsExpressions/meas_expr.h>
 
 namespace gismo
 {
 namespace expr
 {
-
-/*
-  Null expression is a compatibility expression invalid at runtime
-*/
-template<class T>
-class gsNullExpr : public _expr<gsNullExpr<T> >
-{
-public:
-
-    operator const gsFeSpace<T> & () const
-    {
-        static gsFeSpace<T> vv(-1);
-        return vv;
-    }
-
-    typedef T Scalar;
-    gsMatrix<T> eval(const index_t) const { GISMO_ERROR("gsNullExpr"); }
-    inline index_t rows() const { GISMO_ERROR("gsNullExpr"); }
-    inline index_t cols() const { GISMO_ERROR("gsNullExpr"); }
-    void parse(gsExprHelper<T> &) const { }
-
-    const gsFeSpace<T> & rowVar() const { GISMO_ERROR("gsNullExpr"); }
-    const gsFeSpace<T> & colVar() const { GISMO_ERROR("gsNullExpr"); }
-
-    void print(std::ostream &os) const { os << "NullExpr"; }
-
-    static const gsNullExpr & get()
-    {
-        static gsNullExpr o;
-        return o;
-    }
-//private:
-    gsNullExpr() {}
-};
-
-
-template<class E>
-class symbol_expr : public _expr<E>
-{
-public:
-    typedef typename expr_traits<E>::Scalar Scalar;
-
-    friend class gismo::gsExprHelper<Scalar>;
-protected:
-    const gsFunctionSet<Scalar> * m_fs; ///< Evaluation source for this FE variable
-    const gsFuncData<Scalar>    * m_fd; ///< Temporary variable storing flags and evaluation data
-    index_t m_d;                   ///< Dimension of this (scalar or vector) variable
-    bool m_isAcross; ///< true when this expression is evaluated across an interface
-
-public:
-
-    /// Returns whether this expression is evaluated across an interface
-    bool isAcross() const { return m_isAcross; }
-
-    E right() const
-    {
-        E ac(this->derived());
-        ac.m_fs = m_fs;//needed?
-        ac.m_isAcross = true;
-        return ac;
-    }
-
-    E left() const
-    {
-        E ac(this->derived());
-        ac.m_fs = m_fs;
-        ac.m_isAcross = false;
-        return ac;
-    }
-
-    /// Returns the function source
-    const gsFunctionSet<Scalar> & source() const {return *m_fs;}
-
-    /// Returns the function data
-    const gsFuncData<Scalar> & data() const
-    {
-        GISMO_ASSERT(NULL!=m_fd, "FuncData member not registered "<<this<<"/"<< m_fs);
-        return *m_fd;
-    }
-
-    // used by FeSpace, FeVariable, ..
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(*this);
-        this->m_fd->flags |= NEED_VALUE | NEED_ACTIVE;
-    }
-
-    index_t cardinality_impl() const
-    {
-        GISMO_ASSERT(this->data().actives.rows()!=0,"Cardinality depends on the NEED_ACTIVE flag");
-        return m_d * this->data().actives.rows();
-    }
-
-    //public for now due to Bc
-    void setSource(const gsFunctionSet<Scalar> & fs) { m_fs = &fs;}
-
-private:
-    void setData(const gsFuncData<Scalar> & val) { m_fd = &val;}
-    void setDim(index_t _d) { m_d = _d; }
-    void clear() { m_fs = NULL; }
-
-protected:
-    explicit symbol_expr(index_t _d)
-    : m_fs(NULL), m_fd(NULL), m_d(_d), m_isAcross(false) { }
-
-public:
-    bool isValid() const { return NULL!=m_fd && NULL!=m_fs; }
-
-    // component
-    // expr comp(const index_t i) const { return comp_expr<Scalar>(*this,i); }
-    // eval(k).col(i)
-
-    // The evaluation return rows for (basis) functions and columns
-    // for (coordinate) components
-    MatExprType eval(const index_t k) const
-    { return m_fd->values[0].col(k).blockDiag(m_d); } //!!
-    //{ return m_fd->values[0].col(k); }
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-
-    index_t rows() const
-    {
-        GISMO_ASSERT(NULL!=m_fs, "FeVariable: Function source not registered");
-        return m_fs->targetDim();
-    }
-
-    index_t cols() const { return m_d; }
-
-    void print(std::ostream &os) const { os << "u"; }
-
-public:
-
-    /// Returns the vector dimension of the FE variable
-    index_t dim() const { return m_d;}
-
-    /// Returns the target dimension of the FE variable
-    /// before vector-replication
-    index_t targetDim() const { return m_fs->targetDim(); }
-
-    /// Returns the parameter domain dimension the FE variable
-    index_t parDim() const { return m_fs->domainDim(); }
-
-    index_t cSize()  const
-    {
-        GISMO_ASSERT(0!=m_fd->values[0].size(),"Probable error.");
-        return m_fd->values[0].rows();
-    } // coordinate size
-};
-
-/*
-  Column expression
-*/
-template<class E>
-class col_expr : public _expr<col_expr<E> >
-{
-    typename E::Nested_t _c;
-    const index_t _i;
-public:
-    typedef typename E::Scalar Scalar;
-    typedef const col_expr<E> Nested_t;
-
-    enum { Space = E::Space, ScalarValued = 0, ColBlocks = 0 };
-
-    col_expr(const E & c, const index_t i) : _c(c), _i(i) { }
-
-public:
-
-    //ConstColXpr
-    inline MatExprType eval(const index_t k) const { return _c.eval(k).col(_i); }
-
-    index_t rows() const { return _c.rows(); }
-    index_t cols() const { return 1; }
-    void parse(gsExprHelper<Scalar> & evList) const { _c.parse(evList); }
-
-    const gsFeSpace<Scalar> & rowVar() const { return _c.rowVar(); }
-    const gsFeSpace<Scalar> & colVar() const { return _c.colVar(); }
-
-    void print(std::ostream &os) const { os<<_c<<"["<<_i<<"]"; }
-};
 
 /*
   Expression for a constant value
@@ -601,1026 +406,8 @@ public:
     void print(std::ostream &os) const { os<<_c; }
 };
 
-/*
-  Geometry map expression
-*/
-template<class T>
-class gsGeometryMap : public _expr<gsGeometryMap<T> >
-{
-    const gsFunctionSet<T> * m_fs; ///< Evaluation source for this geometry map
-    const gsMapData<T>     * m_fd; ///< Temporary variable storing flags and evaluation data
-    //index_t d, n;
 
-    bool m_isAcross; ///< true when the patch evaluated is across an interface
 
-public:
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    bool isAcross() const { return m_isAcross; }
-
-    gsGeometryMap right() const
-    {
-        gsGeometryMap ac;
-        ac.m_fs = m_fs;
-        ac.m_isAcross = true;
-        return ac;
-    }
-
-    gsGeometryMap left() const
-    {
-        gsGeometryMap ac;
-        ac.m_fs = m_fs;
-        ac.m_isAcross = false;
-        return ac;
-    }
-
-    /// Returns the function source
-    const gsFunctionSet<T> & source() const {return *m_fs;}
-
-    /// Returns the function data
-    const gsMapData<T> & data() const
-    {
-        GISMO_ASSERT(NULL!=m_fd, "gsGeometryMap: invalid data "<< m_fs <<","<<m_fd);
-        return *m_fd;
-    }
-
-    index_t targetDim() const { return m_fs->targetDim();}
-    index_t domainDim() const { return m_fs->domainDim();}
-
-    /// Copy the coefficients of another gsGeometryMap to this one, if they are compatible.
-    void copyCoefs( const gsGeometryMap<T> & other) const
-    {
-        GISMO_ASSERT( dynamic_cast<const gsMultiPatch<T>*>( this->m_fs ), "error");
-        const gsMultiPatch<T> & thisMP  = static_cast<const gsMultiPatch<T>&>(*this->m_fs );
-        GISMO_ASSERT( dynamic_cast<const gsMultiPatch<T>*>( other.m_fs ), "error");
-        const gsMultiPatch<T> & otherMP = static_cast<const gsMultiPatch<T>&>(*other.m_fs );
-        GISMO_ASSERT( (thisMP.domainDim()==otherMP.domainDim())&&
-                      (thisMP.geoDim()==otherMP.geoDim())&&
-                      (thisMP.coefsSize() == otherMP.coefsSize())&&
-                      (thisMP.nPatches()==otherMP.nPatches()),
-                "The geometryMaps are not compatible!");
-
-        // For every patch of the MultiPatch
-        for ( index_t p=0; p < thisMP.nPatches(); p++ )
-        {
-            // Copy coeffs of the other MultiPatch
-            thisMP.patch(p).coefs() = otherMP.patch(p).coefs();
-        }
-
-    }   //end copyCoffs
-
-    void deformBy( const gsFeSolution<T> & displacement) const
-    {
-        const index_t dim = m_fs->domainDim();
-
-        const gsMultiBasis<T> & mb = static_cast<const gsMultiBasis<T>&>(displacement.space().source());
-        const gsMultiPatch<T> & mp = static_cast<const gsMultiPatch<T>&>(*this->m_fs );
-        GISMO_ASSERT( dynamic_cast<const gsMultiBasis<T>*>(&displacement.space().source()), "error");
-        GISMO_ASSERT( dynamic_cast<const gsMultiPatch<T>*>( this->m_fs), "error");
-
-        // For every patch of the MultiPatch
-        for ( size_t p=0; p < mp.nPatches(); p++ )
-        {
-            // Get the patch's coefficients
-            gsMatrix<T> &result = mp.patch(p).coefs();
-
-            // Number of basis functions of patch with index p
-            const index_t sz  = mb[p].size();
-
-            // For all components
-            for (index_t c = 0; c!=dim; c++)
-            {
-                // loop over all basis functions (even the eliminated ones)
-                for (index_t i = 0; i < sz; ++i)
-                {
-                    const int ii = displacement.mapper().index(i, p, c);
-                    if ( displacement.mapper().is_free_index(ii) ) // DoF value is in the defVector
-                    {
-                        result(i,c) += displacement.coefs().at(ii);
-                    }
-                    else
-                    {
-                        result(i,c) += displacement.fixedPart().at( displacement.mapper().global_to_bindex(ii));
-                    }
-                }
-            }
-        }
-    } // end deformBy
-
-public:
-    typedef T Scalar;
-
-    friend class gismo::gsExprHelper<Scalar>;
-
-    void print(std::ostream &os) const { os << "G"; }
-
-    auto eval(const index_t k) const -> decltype(m_fd->values[0].col(k))
-    { return m_fd->values[0].col(k); }
-
-protected:
-
-    gsGeometryMap() : m_fs(NULL), m_fd(NULL), m_isAcross(false) { }
-
-    void setSource(const gsFunctionSet<Scalar> & fs) { m_fs = &fs;}
-    void setData(const gsMapData<Scalar> & val) { m_fd = &val;}
-
-public:
-
-    index_t rows() const { return m_fs->targetDim(); }
-    index_t cols() const { return 1; }
-
-    const gsFeSpace<T> & rowVar() const { return gsNullExpr<T>::get(); }
-    const gsFeSpace<T> & colVar() const { return gsNullExpr<T>::get(); }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(*this);
-        m_fd->flags |= NEED_VALUE;
-    }
-};
-
-// Traits for gsGeometryMap
-template <typename T>  struct expr_traits<gsGeometryMap<T> >
-{
-    typedef T Scalar;
-    typedef const gsGeometryMap<T> Nested_t; // nesting without ref!
-};
-
-/*
-  Expression for the measure of a geometry map
-*/
-template<class T>
-class meas_expr : public _expr<meas_expr<T> >
-{
-    typename gsGeometryMap<T>::Nested_t _G;
-
-public:
-    enum {Space = 0, ScalarValued = 1, ColBlocks = 0};
-
-    typedef T Scalar;
-
-    meas_expr(const gsGeometryMap<T> & G) : _G(G) { }
-
-    T eval(const index_t k) const
-    {
-        return _G.data().measures.at(k);
-    }
-
-    index_t rows() const { return 0; }
-    index_t cols() const { return 0; }
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_G);
-        _G.data().flags |= NEED_MEASURE;
-    }
-
-    const gsFeSpace<T> & rowVar() const { return gsNullExpr<T>::get(); }
-    const gsFeSpace<T> & colVar() const { return gsNullExpr<T>::get(); }
-
-    void print(std::ostream &os) const { os << "meas("; _G.print(os); os <<")"; }
-};
-
-
-/*
-  An element object collecting relevant expressions
-*/
-template<class T>
-class gsFeElement
-{
-    friend class cdiam_expr<T>;
-    const gsExprHelper<T> * m_exprdata;
-
-    gsFeElement(const gsFeElement &);
-public:
-    typedef T Scalar;
-
-    gsFeElement(const gsExprHelper<T> & eh) : m_exprdata(&eh) { }
-
-    bool isValid() const { return nullptr!=m_exprdata; }
-
-    const gsVector<T> & weights() const {return m_exprdata->weights();}
-
-    template<class E> inline
-    integral_expr<E> integral(const _expr<E>& ff) const
-    { return integral_expr<E>(*this,ff); }
-
-    typedef integral_expr<T> AreaRetType;
-    AreaRetType area() const
-    { return integral(_expr<T,true>(1)); }
-
-    typedef integral_expr<meas_expr<T> > PHAreaRetType;
-    /// The diameter of the element on the physical space
-    PHAreaRetType area(const gsGeometryMap<Scalar> & _G) const
-    { return integral(meas_expr<T>(_G)); }
-
-    typedef pow_expr<integral_expr<T> > DiamRetType;
-    /// The diameter of the element (on parameter space)
-    DiamRetType diam() const //-> int(1)^(1/d)
-    { return pow(integral(_expr<T,true>(1)),(T)(1)/(T)(2)); }
-
-    typedef pow_expr<integral_expr<meas_expr<T> > > PHDiamRetType;
-    /// The diameter of the element on the physical space
-    PHDiamRetType diam(const gsGeometryMap<Scalar> & _G) const
-    { return pow(integral(meas_expr<T>(_G)),(T)(1)/(T)(2)); }
-
-    //auto points() const {return point_expr<T>(*this);}
-    //index_t dim() { return di->
-
-    void print(std::ostream &os) const { os << "e"; }
-
-    void parse(gsExprHelper<T> & evList) const
-    {
-        GISMO_ERROR("Call desired member of element expression instead.");
-    }
-};
-
-/**
-   An expression of the element diameter
-*/
-template<class E>
-class integral_expr : public _expr<integral_expr<E> >
-{
-public:
-    //typedef typename E::Scalar Scalar;
-    typedef real_t Scalar;
-    mutable Scalar m_val;
-private:
-    const gsFeElement<Scalar> & _e; ///<Reference to the element
-    typename _expr<E>::Nested_t _ff;
-public:
-    enum {Space= 0, ScalarValued= 1, ColBlocks = 0};
-
-    integral_expr(const gsFeElement<Scalar> & el, const _expr<E> & u)
-    : m_val(-1), _e(el), _ff(u) { }
-
-    const Scalar & eval(const index_t k) const
-    {
-        GISMO_UNUSED(k);
-        GISMO_ENSURE(_e.isValid(), "Element is valid within integrals only.");
-        // if (0==k)
-        {
-            const Scalar * w = _e.weights().data();
-            m_val = (*w) * _ff.val().eval(0);
-            for (index_t j = 1; j != _e.weights().rows(); ++j)
-                m_val += (*(++w)) * _ff.val().eval(j);
-        }
-        return m_val;
-    }
-
-    inline const integral_expr<E> & val() const { return *this; }
-    inline index_t rows() const { return 0; }
-    inline index_t cols() const { return 0; }
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        _ff.parse(evList);
-    }
-
-    const gsFeSpace<Scalar> & rowVar() const { return gsNullExpr<Scalar>::get(); }
-    const gsFeSpace<Scalar> & colVar() const { return gsNullExpr<Scalar>::get(); }
-
-    void print(std::ostream &os) const
-    {
-        os << "integral(";
-        _ff.print(os);
-        os <<")";
-    }
-};
-
-/*
-  template<class T>
-  class parNv_expr : public _expr<parNv_expr<T> >
-  {
-  const gsFeElement<T> & _e;
-  public:
-  typedef T Scalar;
-
-  enum {ScalarValued = 1};
-
-  explicit parNv_expr(const gsFeElement<T> & el) : _e(el) { }
-
-  T eval(const index_t k) const { return _e.m_di->getCellSize(); }
-
-  inline cdiam_expr<T> val() const { return *this; }
-  inline index_t rows() const { return 0; }
-  inline index_t cols() const { return 0; }
-  void parse(gsExprHelper<Scalar> &) const { }
-  const gsFeVariable<T> & rowVar() const { gsNullExpr<Scalar>::get(); }
-  const gsFeVariable<T> & colVar() const { gsNullExpr<Scalar>::get(); }
-
-  void print(std::ostream &os) const
-  { os << "diam(e)"; }
-  };
-*/
-
-template <typename T>
-struct expr_traits<gsFeVariable<T> >
-{
-    typedef T Scalar;
-    typedef const gsFeVariable<T> Nested_t;
-};
-
-/**
-   Expression for finite element variables or PDE coefficient functionals.
-   This can be e.g. a diffusion coefficient, or an isogeometric function.
-*/
-template<class T>
-class gsFeVariable  : public symbol_expr< gsFeVariable<T> >
-{
-    friend class gismo::gsExprHelper<T>;
-    typedef symbol_expr< gsFeVariable<T> > Base;
-protected:
-    explicit gsFeVariable(index_t _d = 1) : Base(_d) { }
-public:
-    enum {Space = 0, ScalarValued = 0, ColBlocks = 0};
-};
-
-template<class T>
-class gsComposition : public symbol_expr< gsComposition<T> >
-{ //comp(f,G)
-    friend class gismo::gsExprHelper<T>;
-    typedef symbol_expr< gsComposition<T> > Base;
-    typename gsGeometryMap<T>::Nested_t _G;
-protected:
-    explicit gsComposition(const gsGeometryMap<T> & G, index_t _d = 1)
-    : Base(_d), _G(G) { }
-public:
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    typename gsMatrix<T>::constColumn
-    eval(const index_t k) const { return this->m_fd->values[0].col(k); }
-
-    const gsGeometryMap<T> & inner() const { return _G;};
-
-    void parse(gsExprHelper<T> & evList) const
-    {
-        //evList.add(_G); //done in gsExprHelper
-        evList.add(*this);
-        this->data().flags |= NEED_VALUE|NEED_ACTIVE;
-        //_G.data().flags  |= NEED_VALUE; //done in gsExprHelper
-    }
-};
-
-
-/**
-   Expression for finite element variable in an isogeometric function
-   space
-
-   This corresponds to an FE variable that
-   belongs to an isogeometric function space
-*/
-template<class T>
-class gsFeSpace :public symbol_expr< gsFeSpace<T> >
-{
-    friend class gsNullExpr<T>;
-
-protected:
-    typedef symbol_expr< gsFeSpace<T> > Base;
-
-    // contains id, mapper, fixedDofs, etc
-    gsFeSpaceData<T> * m_sd;
-
-public:
-    enum{Space = 1, ScalarValued=0, ColBlocks=0};// test space
-
-    typedef const gsFeSpace Nested_t; //no ref
-
-    typedef T Scalar;
-
-    const gsFeSpace<T> & rowVar() const {return *this;}
-
-    gsDofMapper & mapper()
-    {
-        GISMO_ASSERT(NULL!=m_sd, "Space/mapper not properly initialized.");
-        return m_sd->mapper;
-    }
-
-    const gsDofMapper & mapper() const
-    {return const_cast<gsFeSpace*>(this)->mapper();}
-
-    inline const gsMatrix<T> & fixedPart() const {return m_sd->fixedDofs;}
-    gsMatrix<T> & fixedPart() {return m_sd->fixedDofs;}
-
-    index_t   id() const { return (m_sd ? m_sd->id : -101); }
-    void setSpaceData(gsFeSpaceData<T>& sd) {m_sd = &sd;}
-
-    index_t   interfaceCont() const {return m_sd->cont;}
-    index_t & setInterfaceCont(const index_t _r) const
-    {
-        GISMO_ASSERT(_r>-2 && _r<1, "Invalid or not implemented (r="<<_r<<").");
-        return m_sd->cont = _r;
-    }
-
-    gsFeSolution<T> function(const gsMatrix<T>& solVector) const
-    { return gsFeSolution<T>(*this); }
-
-    void getCoeffs(const gsMatrix<T>& solVector, gsMatrix<T> & result,
-                   const index_t p = 0) const
-    {
-        const index_t dim = this->dim();
-
-        const gsMultiBasis<T> & mb = static_cast<const gsMultiBasis<T>&>(this->source());
-        GISMO_ASSERT( dynamic_cast<const gsMultiBasis<T>*>(&this->source()), "error");
-
-        // Reconstruct solution coefficients on patch p
-        const index_t sz  = mb[p].size();
-        result.resize(sz, dim!=1 ? dim : solVector.cols()); // (!)
-
-        for (index_t c = 0; c!=dim; c++) // for all components
-        {
-            // loop over all basis functions (even the eliminated ones)
-            for (index_t i = 0; i < sz; ++i)
-            {
-                const int ii = m_sd->mapper.index(i, p, c);
-                if ( m_sd->mapper.is_free_index(ii) ) // DoF value is in the solVector
-                  for(index_t s = 0; s != solVector.cols(); ++s )
-                    result(i,c+s) = solVector(ii,s); //assume dim==1 xor solVector.cols()==1
-                else // eliminated DoF: fill with Dirichlet data
-                    result(i,c) =  m_sd->fixedDofs.at( m_sd->mapper.global_to_bindex(ii) );
-            }
-        }
-    }
-
-    // space restrictTo(boundaries);
-    // space restrictTo(bcRefList domain);
-
-    void setupMapper(gsDofMapper dofsMapper) const
-    {
-        GISMO_ASSERT( dofsMapper.isFinalized(), "The provided dof-mapper is not finalized.");
-        GISMO_ASSERT( dofsMapper.mapSize()==static_cast<size_t>(this->source().size()*dofsMapper.numComponents()), "The dof-mapper is not consistent: mapSize()="<<dofsMapper.mapSize()<<"!="<<static_cast<size_t>(this->source().size())<<"=this->source().size()");
-        m_sd->mapper = give(dofsMapper);
-    }
-
-    void setup(const index_t _icont = -1) const
-    {
-        this->setInterfaceCont(_icont);
-        m_sd->mapper = gsDofMapper();
-
-        if (const gsMultiBasis<T> * mb =
-            dynamic_cast<const gsMultiBasis<T>*>(&this->source()) )
-        {
-            m_sd->mapper = gsDofMapper(*mb, this->dim() );
-            //m_mapper.init(*mb, this->dim()); //bug
-            if ( 0==this->interfaceCont() ) // Conforming boundaries ?
-            {
-                for ( gsBoxTopology::const_iiterator it = mb->topology().iBegin();
-                      it != mb->topology().iEnd(); ++it )
-                {
-                    mb->matchInterface(*it, m_sd->mapper);
-                }
-            }
-        }
-
-        if (const gsMappedBasis<2,T> * mb =
-            dynamic_cast<const gsMappedBasis<2,T>*>(&this->source()) )
-        {
-            m_sd->mapper.setIdentity(mb->nPatches(), mb->size() , this->dim());
-        }
-
-        m_sd->mapper.finalize();
-    }
-
-    void setup(const gsBoundaryConditions<T> & bc, const index_t dir_values,
-               const index_t _icont = -1) const
-    {
-        this->setInterfaceCont(_icont);
-        m_sd->mapper = gsDofMapper();
-        const index_t dim = this->dim();
-        const gsMultiBasis<T> *mb = dynamic_cast<const gsMultiBasis<T> *>(&this->source());
-        if (mb != nullptr)
-        {
-            m_sd->mapper = gsDofMapper(*mb, this->dim());
-            //m_mapper.init(*mb, this->dim()); //bug
-            if (0 == this->interfaceCont()) // Conforming boundaries ?
-            {
-                for (gsBoxTopology::const_iiterator it = mb->topology().iBegin();
-                     it != mb->topology().iEnd(); ++it) {
-                    if ( it->type() != interaction::contact ) // If the interface type is 'contact' ignore it.
-                        mb->matchInterface(*it, m_sd->mapper);
-                }
-            }
-
-            // Strong Dirichlet conditions
-            gsMatrix<index_t> bnd;
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Dirichlet"); it != bc.end("Dirichlet"); ++it)
-            {
-                const index_t cc = it->unkComponent();
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mb->basis(it->ps.patch).boundary(it->ps.side());
-                m_sd->mapper.markBoundary(it->ps.patch, bnd, cc);
-            }
-            // Clamped boundary condition (per DoF)
-            gsMatrix<index_t> bnd1;
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Clamped"); it != bc.end("Clamped"); ++it)
-            {
-                const index_t cc = it->unkComponent();
-
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mb->basis(it->ps.patch).boundaryOffset(it->ps.side(), 0);
-                bnd1 = mb->basis(it->ps.patch).boundaryOffset(it->ps.side(), 1);
-                // Cast to tensor b-spline basis
-                if (!it->ps.parameter())
-                        bnd.swap(bnd1);
-                for (index_t c = 0; c!=dim; c++) // for all components
-                {
-                    if (c==cc || cc==-1 )
-                        for (index_t k = 0; k < bnd.size(); ++k)
-                            m_sd->mapper.matchDof(it->ps.patch, (bnd)(k, 0),
-                                                  it->ps.patch, (bnd1)(k, 0), c);
-                }
-
-            }
-
-            // Collapsed
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Collapsed"); it != bc.end("Collapsed"); ++it)
-            {
-                const index_t cc = it->unkComponent();
-
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mb->basis(it->ps.patch).boundary(it->ps.side());
-
-                // match all DoFs to the first one of the side
-                for (index_t c = 0; c!=dim; c++) // for all components
-                {
-                    if (c==cc || cc==-1)
-                        for (index_t k = 0; k < bnd.size() - 1; ++k)
-                            m_sd->mapper.matchDof(it->ps.patch, (bnd)(0, 0),
-                                                  it->ps.patch, (bnd)(k + 1, 0), c);
-                }
-            }
-
-            // Coupled
-            for (typename gsBoundaryConditions<T>::const_cpliterator
-                     it = bc.coupledBegin(); it != bc.coupledEnd(); ++it)
-            {
-                const index_t cc = it->component;
-
-                GISMO_ASSERT(static_cast<size_t>(it->ifc.first().patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-                GISMO_ASSERT(static_cast<size_t>(it->ifc.second().patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-
-                bnd = mb->basis(it->ifc.first().patch).boundary(it->ifc.first().side());
-                bnd1 = mb->basis(it->ifc.second().patch).boundary(it->ifc.second().side());
-
-                // match all DoFs to the first one of the side
-                for (index_t c = 0; c!=dim; c++) // for all components
-                {
-                    if (c==cc || cc==-1)
-                    {
-                        for (index_t k = 0; k < bnd.size() - 1; ++k)
-                            m_sd->mapper.matchDof(it->ifc.first() .patch, (bnd)(0, 0),
-                                                  it->ifc.first() .patch, (bnd)(k + 1, 0), c);
-                        for (index_t k = 0; k < bnd1.size(); ++k)
-                            m_sd->mapper.matchDof(it->ifc.first() .patch, (bnd)(0, 0),
-                                                  it->ifc.second().patch, (bnd1)(k, 0), c);
-                    }
-                }
-            }
-
-            // corners
-            for (typename gsBoundaryConditions<T>::const_citerator
-                     it = bc.cornerBegin(); it != bc.cornerEnd(); ++it)
-            {
-                for (index_t r = 0; r!=this->dim(); ++r)
-                {
-                    if (it->component!=-1 && r!=it->component) continue;
-
-                    //assumes (unk == -1 || it->unknown == unk)
-                    GISMO_ASSERT(static_cast<size_t>(it->patch) < mb->nBases(),
-                                 "Problem: a corner boundary condition is set on a patch id which does not exist.");
-                    m_sd->mapper.eliminateDof(mb->basis(it->patch).functionAtCorner(it->corner),
-                                              it->patch, it->component);
-                }
-            }
-
-        } else if (const gsBasis<T> *b =
-                   dynamic_cast<const gsBasis<T> *>(&this->source()))
-        {
-            m_sd->mapper = gsDofMapper(*b, this->dim() );
-            gsMatrix<index_t> bnd;
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Dirichlet"); it != bc.end("Dirichlet"); ++it) {
-                GISMO_ASSERT(it->ps.patch == 0,
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = b->boundary(it->ps.side());
-                m_sd->mapper.markBoundary(0, bnd, it->unkComponent());
-            }
-
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Clamped"); it != bc.end("Clamped"); ++it) {
-                GISMO_ASSERT(it->ps.patch == 0,
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = b->boundary(it->ps.side());
-                //const index_t cc = it->unkComponent();
-                // m_sd->mapper.markBoundary(0, bnd, 0);
-            }
-
-            m_sd->mapper = gsDofMapper(*b);
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Collapsed"); it != bc.end("Collapsed"); ++it) {
-                GISMO_ASSERT(it->ps.patch == 0,
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = b->boundary(it->ps.side());
-                //const index_t cc = it->unkComponent();
-                // m_sd->mapper.markBoundary(0, bnd, 0);
-            }
-        } else if (const gsMappedBasis<2, T> *mapb =
-                   dynamic_cast<const gsMappedBasis<2, T> *>(&this->source()))
-        {
-            m_sd->mapper.setIdentity(mapb->nPatches(), mapb->size(), this->dim());
-
-            if (0 == this->interfaceCont()) // C^0 matching interface
-            {
-                gsMatrix<index_t> int1, int2;
-                for (gsBoxTopology::const_iiterator it = mapb->getTopol().iBegin();
-                     it != mapb->getTopol().iEnd(); ++it) {
-                    int1 = mapb->basis(it->first().patch).boundaryOffset(it->first().side(), 0);
-                    int2 = mapb->basis(it->second().patch).boundaryOffset(it->second().side(), 0);
-
-                    m_sd->mapper.matchDofs(it->first().patch, int1, it->second().patch, int2);
-                }
-            }
-            if (1 == this->interfaceCont()) // C^1 matching interface
-            {
-                GISMO_ERROR("Boundary offset function is not implemented for gsMappedBasis in general.");
-            }
-
-            gsMatrix<index_t> bnd;
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Dirichlet"); it != bc.end("Dirichlet"); ++it) {
-                const index_t cc = it->unkComponent();
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mapb->basis(it->ps.patch).boundary(it->ps.side());
-                m_sd->mapper.markBoundary(it->ps.patch, bnd, cc);
-            }
-
-            // Clamped boundary condition (per DoF)
-            gsMatrix<index_t> bnd1;
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Clamped"); it != bc.end("Clamped"); ++it) {
-                const index_t cc = it->unkComponent();
-
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mapb->basis(it->ps.patch).boundaryOffset(it->ps.side(), 0);
-                bnd1 = mapb->basis(it->ps.patch).boundaryOffset(it->ps.side(), 1);
-
-                // Cast to tensor b-spline basis
-                if (mapb != NULL) // clamp adjacent dofs
-                {
-                    if (!it->ps.parameter())
-                        bnd.swap(bnd1);
-                    for (index_t c = 0; c!=dim; c++) // for all components
-                    {
-                        if (c==cc || cc==-1 )
-                            for (index_t k = 0; k < bnd.size() - 1; ++k)
-                                m_sd->mapper.matchDof(  it->ps.patch, (bnd)(k, 0),
-                                                        it->ps.patch, (bnd1)(k, 0), c);
-                    }
-                } else
-                    gsWarn << "Unable to apply clamped condition.\n";
-            }
-
-            // COLLAPSED
-            for (typename gsBoundaryConditions<T>::const_iterator
-                     it = bc.begin("Collapsed"); it != bc.end("Collapsed"); ++it) {
-                const index_t cc = it->unkComponent();
-
-                GISMO_ASSERT(static_cast<size_t>(it->ps.patch) < this->mapper().numPatches(),
-                             "Problem: a boundary condition is set on a patch id which does not exist.");
-
-                bnd = mapb->basis(it->ps.patch).boundary(it->ps.side());
-
-                // Cast to tensor b-spline basis
-                if (mapb != NULL) // clamp adjacent dofs
-                {
-                    // match all DoFs to the first one of the side
-                    for (index_t c = 0; c!=dim; c++) // for all components
-                    {
-                        if (c==cc || cc==-1)
-                            for (index_t k = 0; k < bnd.size() - 1; ++k)
-                                m_sd->mapper.matchDof(it->ps.patch, (bnd)(0, 0),
-                                                      it->ps.patch, (bnd)(k + 1, 0), c);
-                    }
-                }
-            }
-
-            // corners
-            for (typename gsBoundaryConditions<T>::const_citerator
-                     it = bc.cornerBegin(); it != bc.cornerEnd(); ++it)
-            {
-                //assumes (unk == -1 || it->unknown == unk)
-                GISMO_ASSERT(it->patch < mapb->nPieces(),
-                             "Problem: a corner boundary condition is set on a patch id which does not exist.");
-                m_sd->mapper.eliminateDof(mapb->basis(it->patch).functionAtCorner(it->corner), it->patch, it->component);
-            }
-        } else
-        {
-            GISMO_ASSERT(0 == bc.size(), "Problem: BCs are ignored.");
-            m_sd->mapper.setIdentity(this->source().nPieces(), this->source().size());
-        }
-
-        m_sd->mapper.finalize();
-
-        // Compute Dirichlet node values
-        gsDirichletValues(bc, dir_values, *this);
-    }
-
-    void print(std::ostream &os) const { os << "u"; }
-
-protected:
-    friend class gismo::gsExprHelper<Scalar>;
-    friend class symbol_expr<gsFeSpace>;
-    explicit gsFeSpace(index_t _d = 1) : Base(_d), m_sd(nullptr) { }
-};
-
-template<class T> inline bool
-operator== (const gsFeSpace<T> & a, const gsFeSpace<T> & b)
-{ return a.id()== b.id() && a.isAcross()==b.isAcross(); }
-
-/*
-  Expression representing a function given by a vector of
-  coefficients in a gsFeSpace.
-
-  Typically it used for accessing the solution of a boundary-value
-  problem.
-*/
-template<class T>
-class gsFeSolution : public _expr<gsFeSolution<T> >
-{
-protected:
-    const gsFeSpace<T> _u;
-    gsMatrix<T> * _Sv; ///< Pointer to a coefficient vector
-    bool m_isAcross; ///< true when this expression is evaluated across an interface
-
-public:
-    typedef T Scalar;
-    enum {Space = 0, ScalarValued= 0, ColBlocks= 0};
-
-    bool isAcross() const { return m_isAcross; }
-
-    gsFeSolution right() const
-    {
-        gsFeSolution ac(*this);
-        ac.m_isAcross = true;
-        return ac;
-    }
-
-    gsFeSolution left() const { return gsFeSolution(*this); }
-
-    explicit gsFeSolution(const gsFeSpace<T> & u) : _u(u), _Sv(NULL) { }
-
-    gsFeSolution(const gsFeSpace<T> & u, gsMatrix<T> & Sv) : _u(u), _Sv(&Sv) { }
-
-    const gsFeSpace<T> & space() const {return _u;};
-
-    mutable gsMatrix<T> res;
-    const gsMatrix<T> & eval(index_t k) const
-    {
-        GISMO_ASSERT(check(), "Invalid state in gsFeSolution");
-        const gsDofMapper & map = _u.mapper();
-        auto & act = _u.data().actives.col(1 == _u.data().actives.cols() ? 0:k );
-        res.setZero(_u.dim(), 1);
-        for (index_t c = 0; c!=_u.dim(); c++) // for all components
-        {
-            for (index_t i = 0; i!=_u.data().actives.rows(); ++i)
-            {
-                const index_t ii = map.index( act[i], _u.data().patchId, c);
-                if ( map.is_free_index(ii) ) // DoF value is in the solVector
-                    res.at(c) += _Sv->at(ii) * _u.data().values[0](i,k);
-                else
-                    res.at(c) += _u.data().values[0](i,k) *
-                        _u.fixedPart().at( map.global_to_bindex(ii) );
-            }
-        }
-        return res;
-    }
-
-    // Performs validity checks for the solution object
-    bool check() const
-    {
-        if ( _Sv->size()!=_u.mapper().freeSize() )
-        {
-            gsWarn<< "The solution vector has wrong dimensions: "
-                  <<_Sv->size()<<" != "<<_u.mapper().freeSize() <<". ";
-            return false;
-        }
-        if((size_t)_u.source().size()*dim()!=_u.mapper().mapSize())
-        {
-            gsWarn<< "The solution space is inconsistent: "
-                  <<_u.source().size()*dim()<<" != "<<_u.mapper().mapSize()<<". ";
-            return false;
-        }
-        return true;
-    }
-    
-    //template<class U>
-    //linearComb(U & ie){ sum up ie[_u] times the _Sv  }
-    // ie.eval(k), _u.data().actives(), fixedPart() - see lapl_expr
-
-    const gsFeSpace<Scalar> & rowVar() const {return gsNullExpr<Scalar>::get();}
-    const gsFeSpace<Scalar> & colVar() const {return gsNullExpr<Scalar>::get();}
-    index_t rows() const {return _u.dim(); }
-
-    static index_t cols() {return 1; }
-
-    void parse(gsExprHelper<Scalar> & evList) const
-    {
-        evList.add(_u);
-        _u.data().flags |= NEED_VALUE | NEED_ACTIVE;
-    }
-
-    void print(std::ostream &os) const { os << "s"; }
-
-public:
-    index_t dim() const { return _u.dim();}
-
-    index_t parDim() const
-    { return _u.source().domainDim(); }
-
-    //gsDofMapper & mapper() {return _u.mapper();}
-    const gsDofMapper & mapper() const {return _u.mapper();}
-
-    inline const gsMatrix<T> & fixedPart() const {return _u.fixedPart();}
-    gsMatrix<T> & fixedPart() {return _u.fixedPart();}
-
-    //gsFuncData<T> & data() {return _u.data();}
-    const gsFuncData<T> & data() const {return _u.data();}
-
-    void setSolutionVector(gsMatrix<T>& solVector)
-    { _Sv = & solVector; }
-
-    /// @brief Sets all coefficients of the solution vector that belong to
-    ///    patch \a p , and refer to the specified \a component equal to \a value.
-    /// @param component The index of the component to be set.
-    /// @param value The value that the coefficients will be set to.
-    /// @param patch The index of the patch whose coefficients will be set. By default all patches are affected,
-    void setComponent(index_t component, real_t value, index_t patch=-1)
-    {
-        gsMatrix<T> & solVector = *_Sv;
-        const gsDofMapper & mapper = _u.mapper();
-
-        index_t patchStart, patchEnd;
-        if (patch==-1){
-            patchStart = 0;
-            patchEnd   = _u.mapper().numPatches();
-        }
-        else{
-            patchStart = patch;
-            patchEnd   = patch + 1;
-        }
-
-        for (index_t p=patchStart; p!=patchEnd; ++p)
-        {
-            for (size_t i = 0; i != mapper.patchSize(p, component); ++i)
-            {
-                const index_t ii = mapper.index(i, p, component);
-                if ( mapper.is_free_index(ii) ) // DoF value is in the solVector
-                    solVector.at(ii) = value;
-            }
-        }
-    }
-
-    const gsMatrix<T> & coefs() const { return *_Sv; }
-    //gsMatrix<T> & coefs() { return *_Sv; } // wd4702 ?
-
-    //const gsMatrix<T> & coefs(component, patch) const { return *_Sv; }
-
-    /// val: perturbation value, j: global index, p: patch
-    void perturbLocal(T val, index_t j, index_t p = 0)
-    {
-        GISMO_UNUSED(p);
-        // GISMO_ASSERT(1==_u.data().actives.cols(), "Single actives expected");
-        //if (_u.mapper().is_free_index(j) )
-        //{
-            GISMO_ASSERT(j<_Sv->size(), "Solution vector is not initialized/allocated, sz="<<_Sv->size() );
-            _Sv->at(j) += val;
-            //}
-        //else
-        //    _u.fixedPart().at( _u.mapper().global_to_bindex(j) ) += val;
-    }
-
-    /// Extract the coefficients of piece \a p
-    void extract(gsMatrix<T> & result, const index_t p = 0) const
-    { _u.getCoeffs(*_Sv, result, p); }
-
-    /// Extracts ALL the coefficients in a solution vector; including
-    /// coupled and boundary DoFs
-    void extractFull(gsMatrix<T> & result) const
-    {
-        index_t offset;
-        const index_t dim = _u.dim();
-        const size_t totalSz = _u.mapper().mapSize();
-        result.resize(totalSz, 1);
-        for (size_t p=0; p!=_u.mapper().numPatches(); ++p)
-        {
-            offset = _u.mapper().offset(p);
-            // Reconstruct solution coefficients on patch p
-
-            for (index_t c = 0; c!=dim; c++) // for all components
-            {
-                const index_t sz  = _u.mapper().patchSize(p,c);
-
-                // loop over all basis functions (even the eliminated ones)
-                for (index_t i = 0; i < sz; ++i)
-                {
-                    //gsDebugVar(i);
-                    const int ii = _u.mapper().index(i, p, c);
-                    //gsDebugVar(ii);
-                    if ( _u.mapper().is_free_index(ii) ) // DoF value is in the solVector
-                    {
-                        result(i+offset,0) = _Sv->at(ii);
-                    }
-                    else // eliminated DoF: fill with Dirichlet data
-                        result(i+offset,0) =  _u.fixedPart().at( _u.mapper().global_to_bindex(ii) );
-                }
-                offset += sz;
-            }
-        }
-    }
-
-    /// Extract this variable as a multipatch object
-    void extract(gsMultiPatch<T> & result) const
-    {
-        result.clear();
-
-        if( const gsMultiBasis<T>* basis = dynamic_cast<const gsMultiBasis<T>* >(&_u.source()) )
-            for (size_t i = 0; i != basis->nBases(); ++i)
-            {
-                memory::unique_ptr<gsGeometry<T> > p(this->extractPiece(i));
-                result.addPatch(*p);
-            }
-    }
-
-    /// Extract this variable as a gsMappedSpline object
-    void extract(gsMappedSpline<2,T> & result) const
-    {
-        if( const gsMappedBasis<2,T>* basis = dynamic_cast<const gsMappedBasis<2,T>* >(&_u.source()) )
-        {
-            gsMatrix<T> coefs;
-            this->extractFull(coefs);
-            coefs.resize(coefs.rows()/_u.dim(),_u.dim());
-            result.init(*basis,coefs);
-        }
-    }
-
-    /// Extract the piece \a p as a gsGeometry pointer
-    memory::unique_ptr<gsGeometry<T> > extractPiece(const index_t p) const
-    {
-        if ( const gsBasis<T> * b = dynamic_cast<const gsBasis<T>*>(&_u.source().piece(p)) )
-        {
-            gsMatrix<T> cf;
-            extract(cf, p);
-            return b->makeGeometry(give(cf));
-        }
-        GISMO_ERROR("gsFeSolution: Extraction error");
-    }
-
-    // insert g-coefficients to the solution vector
-    void insert(const gsGeometry<T> & g, const index_t p = 0) const
-    {
-        insert(g.coefs(), p);
-    }
-
-    // insert g-coefficients to the solution vector
-    void insert(const gsMatrix<T> & cf, const index_t p = 0) const
-    {
-        gsMatrix<T> & sol = *_Sv;
-        //gsMatrix<T> & fixedPart = _u.fixedPart();
-        const gsDofMapper & mapper = _u.mapper();
-        for (index_t c = 0; c!=_u.dim(); c++) // for all components
-        {
-            for (index_t i = 0; i != cf.rows(); ++i)
-            {
-                const index_t ii = mapper.index(i, p, c);
-                if ( mapper.is_free_index(ii) ) // DoF value is in the solVector
-                    sol.at(ii) = cf(i, c);
-                /*
-                  else
-                  {
-                  fixedPart.row(m_sd->mapper.global_to_bindex(ii)) = cf.row(i);
-                  }
-                */
-            }
-        }
-    }
-};
 
 /*
   Expression for the transpose of an expression
