+++ src/gsNurbs/gsKnotVector.h
@@ -165,7 +165,7 @@ public: // iterator ends
         return new gsKnotDomainIterator<T>(*this);
     }
 
-    domainIter beginBdr(const boxSide   bs) const override
+    domainIter beginBdr(const boxSide /* bs */) const override
     {
         return new gsKnotDomainIterator<T>(*this);
     }
@@ -175,7 +175,7 @@ public: // iterator ends
         return new gsKnotDomainIterator<T>(*this,false);
     }
 
-    domainIter endBdr(const boxSide   bs) const override
+    domainIter endBdr(const boxSide /* bs */) const override
     {
         return new gsKnotDomainIterator<T>(*this,false);
     }
@@ -739,6 +739,7 @@ public: // things required by gsKnotVector
     /// Returns the degree of the knot vector.
     short_t degree(short_t i = 0) const override //overload from gsDomain
     {
+        GISMO_UNUSED(i);
         return m_deg;
     }
